Formatting of solidity files

+++ packages/contracts/contracts/votings/whitelist/WhitelistVoting.sol
@@ -1,6 +1,4 @@
-/*
- * SPDX-License-Identifier:    MIT
- */
+// SPDX-License-Identifier: MIT
 
 pragma solidity 0.8.10;
 
@@ -98,18 +96,18 @@ contract WhitelistVoting is MajorityVoting {
         VoterState _choice
     ) external override returns (uint256 voteId) {
         uint64 snapshotBlock = getBlockNumber64() - 1;
-        
-        if(!isUserWhitelisted(_msgSender(), snapshotBlock)) {
+
+        if (!isUserWhitelisted(_msgSender(), snapshotBlock)) {
             revert VoteCreationForbidden(_msgSender());
         }
-        
+
         // calculate start and end time for the vote
         uint64 currentTimestamp = getTimestamp64();
 
         if (_startDate == 0) _startDate = currentTimestamp;
         if (_endDate == 0) _endDate = _startDate + minDuration;
 
-        if(_endDate - _startDate <  minDuration || _startDate < currentTimestamp)
+        if (_endDate - _startDate < minDuration || _startDate < currentTimestamp)
             revert VoteTimesForbidden({
                 current: currentTimestamp,
                 start: _startDate,
@@ -181,66 +179,54 @@ contract WhitelistVoting is MajorityVoting {
             _execute(_voteId);
         }
     }
-   
-    /**
-    *  @dev Tells whether user is whitelisted at specific block or past it.
-    *  @param account user address
-    *  @param blockNumber block number for which it checks if user is whitelisted
-    */
-    function isUserWhitelisted(
-        address account, 
-        uint256 blockNumber
-    ) public view returns (bool) {
-        if(blockNumber == 0) blockNumber = getBlockNumber64() - 1;
-        
+
+    /// @dev Tells whether user is whitelisted at specific block or past it.
+    /// @param account user address
+    /// @param blockNumber block number for which it checks if user is whitelisted
+    function isUserWhitelisted(address account, uint256 blockNumber) public view returns (bool) {
+        if (blockNumber == 0) blockNumber = getBlockNumber64() - 1;
+
         return _checkpoints[account].getAtBlock(blockNumber) == 1;
     }
 
-    /**
-    *  @dev returns total count of users that are whitelisted at specific block
-    *  @param blockNumber specific block to get count from
-    *  @return count of users that are whitelisted blockNumber or prior to it.
-    */
-    function whitelistedUserCount(
-        uint256 blockNumber
-    ) public view returns (uint256) {
-        if(blockNumber == 0) blockNumber = getBlockNumber64() - 1;
-        
+    /// @dev returns total count of users that are whitelisted at specific block
+    /// @param blockNumber specific block to get count from
+    /// @return count of users that are whitelisted blockNumber or prior to it.
+    function whitelistedUserCount(uint256 blockNumber) public view returns (uint256) {
+        if (blockNumber == 0) blockNumber = getBlockNumber64() - 1;
+
         return _totalCheckpoints.getAtBlock(blockNumber);
     }
 
-    /**
-     * @dev Internal function to check if a voter can participate on a vote. It assumes the queried vote exists.
-     * @param _voteId The voteId
-     * @param _voter the address of the voter to check
-     * @return True if the given voter can participate a certain vote, false otherwise
-     */
+    /// @dev Internal function to check if a voter can participate on a vote. It assumes the queried vote exists.
+    /// @param _voteId The voteId
+    /// @param _voter the address of the voter to check
+    /// @return True if the given voter can participate a certain vote, false otherwise
     function _canVote(uint256 _voteId, address _voter) internal view override returns (bool) {
         Vote storage vote_ = votes[_voteId];
         return _isVoteOpen(vote_) && isUserWhitelisted(_voter, vote_.snapshotBlock);
     }
 
-    /**
-    *  @dev Adds or removes users from whitelist
-    *  @param _users user addresses
-    *  @param _enabled whether to add or remove from whitelist
-    */
-    function _whitelistUsers(
-        address[] calldata _users, 
-        bool _enabled
-    ) internal {        
+    /// @dev Adds or removes users from whitelist
+    /// @param _users user addresses
+    /// @param _enabled whether to add or remove from whitelist
+    function _whitelistUsers(address[] calldata _users, bool _enabled) internal {
         _totalCheckpoints.push(_enabled ? _add : _sub, _users.length);
-        
-        for(uint i = 0; i < _users.length; i++) {
+
+        for (uint256 i = 0; i < _users.length; i++) {
             _checkpoints[_users[i]].push(_enabled ? 1 : 0);
         }
     }
 
-    function _add(uint256 a, uint256 b) private pure returns(uint256) {
-        unchecked { return a + b; }
+    function _add(uint256 a, uint256 b) private pure returns (uint256) {
+        unchecked {
+            return a + b;
+        }
     }
 
-    function _sub(uint256 a, uint256 b) private pure returns(uint256) {
-        unchecked { return a - b; }
+    function _sub(uint256 a, uint256 b) private pure returns (uint256) {
+        unchecked {
+            return a - b;
+        }
     }
 }
